Change trie from 6-ary to 36-ary
This makes each represented byte 3x larger (from 48B to 144B) but cuts
in half the number of memory read operations to check membership. This
almost cancels out the speed penalty of using the __global memory space.
In contrast with the 150 terms that could be held in the __constant
cache, with 1GB of memory you can hold (in the worst case) 771 thousand
terms.